Refactor ActiveVersioning errors
