+++ lib/active_versioning/errors/incompatible_version.rb
@@ -1,0 +1,12 @@
+module ActiveVersioning
+  module Errors
+    class IncompatibleVersion < StandardError
+      attr_reader :record, :version
+
+      def initialize(record, version)
+        @record  = record
+        @version = version
+      end
+    end
+  end
+end
